# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FUTRSEC - Cybersecurity Learning + Mentorship + AI + Placement Ecosystem API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /consent:
    get:
      operationId: getConsentStatus
      tags: [consent]
      summary: Get current user consent status
      responses:
        "200":
          description: Consent status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentStatus"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    post:
      operationId: captureConsent
      tags: [consent]
      summary: Capture user consent (called during onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentCaptureInput"
      responses:
        "200":
          description: Consent recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentStatus"
        "400":
          description: Invalid consent data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /consent/history:
    get:
      operationId: getConsentHistory
      tags: [consent]
      summary: Get immutable consent audit trail for current user
      responses:
        "200":
          description: Consent history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsentHistoryEntry"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /consent/withdraw:
    post:
      operationId: withdrawConsent
      tags: [consent]
      summary: Withdraw a specific consent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentWithdrawInput"
      responses:
        "200":
          description: Consent withdrawn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentStatus"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /dpdp/download-request:
    post:
      operationId: requestDataDownload
      tags: [dpdp]
      summary: Request personal data download (DPDP right)
      responses:
        "202":
          description: Download request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /dpdp/delete-request:
    post:
      operationId: requestDataDeletion
      tags: [dpdp]
      summary: Request account and personal data deletion (DPDP right)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataDeletionInput"
      responses:
        "202":
          description: Deletion request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /dpdp/correction-request:
    post:
      operationId: requestDataCorrection
      tags: [dpdp]
      summary: Request correction of personal data (DPDP right)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataCorrectionInput"
      responses:
        "202":
          description: Correction request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /dpdp/requests:
    get:
      operationId: listDataRequests
      tags: [dpdp]
      summary: List current user's DPDP data requests
      responses:
        "200":
          description: List of data requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ───────────────────────────── TRACKS ──────────────────────────────
...
  /tracks/{id}:
    get:
      operationId: getTrack
      tags: [tracks]
      summary: Get a specific track with its modules
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Track detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackDetail"
        "404":
          description: Track not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ───────────────────────────── ASSESSMENTS ──────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: ["string", "null"]

    # ── Auth ──
...
    ConsentStatus:
      type: object
      required: [userId, marketing, analytics, dataProcessing, thirdParty, updatedAt]
      properties:
        userId:
          type: integer
        marketing:
          type: boolean
        analytics:
          type: boolean
        dataProcessing:
          type: boolean
        thirdParty:
          type: boolean
        updatedAt:
          type: string
          format: date-time
    ConsentCaptureInput:
      type: object
      required: [dataProcessing]
      properties:
        marketing:
          type: boolean
        analytics:
          type: boolean
        dataProcessing:
          type: boolean
        thirdParty:
          type: boolean
    ConsentWithdrawInput:
      type: object
      required: [consentType]
      properties:
        consentType:
          type: string
          enum: [marketing, analytics, dataProcessing, thirdParty]
        reason:
          type: string
    ConsentHistoryEntry:
      type: object
      required: [id, userId, consentType, action, ipAddress, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        consentType:
          type: string
        action:
          type: string
          enum: [granted, withdrawn, updated]
        ipAddress:
          type: string
        userAgent:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    DataRequest:
      type: object
      required: [id, userId, type, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [download, deletion, correction]
        status:
          type: string
          enum: [pending, processing, completed, rejected]
        notes:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    DataDeletionInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    DataCorrectionInput:
      type: object
      required: [field, currentValue, requestedValue]
      properties:
        field:
          type: string
        currentValue:
          type: string
        requestedValue:
          type: string
        supportingNote:
          type: string

    # ── Tracks ──
...
    TrackDetail:
      type: object
      required: [id, name, slug, domain, description, difficulty, durationWeeks, modules]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        domain:
          type: string
        description:
          type: string
        difficulty:
          type: string
        durationWeeks:
          type: integer
        accentColor:
          type: string
        modules:
          type: array
          items:
            $ref: "#/components/schemas/LearningModule"
    LearningModule:
      type: object
      required: [id, title, order, lessonCount]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        lessonCount:
          type: integer

    # ── Assessments ──
...